Mass update
If data directory is not in jar file, it is created.
Bug: tasks.txt folder added inside data folder, which is redundant

Remove Duke2 which is a dead code
Add Ui.png
Remove dead print statements
