+++ src/main/java/seedu/deku/DekuException.java
@@ -1,0 +1,9 @@
+package seedu.deku;
+
+public class DekuException extends Exception {
+
+    public DekuException(String message) {
+        super(message);
+    }
+
+}
