+++ src/main/java/seedu/deku/MainWindow.java
@@ -1,0 +1,60 @@
+package seedu.deku;
+
+import javafx.application.Platform;
+import javafx.fxml.FXML;
+import javafx.scene.control.Button;
+import javafx.scene.control.ScrollPane;
+import javafx.scene.control.TextField;
+import javafx.scene.image.Image;
+import javafx.scene.layout.AnchorPane;
+import javafx.scene.layout.VBox;
+
+import java.io.IOException;
+
+/**
+ * Controller for MainWindow. Provides the layout for the other controls.
+ */
+public class MainWindow extends AnchorPane {
+    @FXML
+    private ScrollPane scrollPane;
+    @FXML
+    private VBox dialogContainer;
+    @FXML
+    private TextField userInput;
+    @FXML
+    private Button sendButton;
+
+    private Deku deku;
+
+    private Image userImage = new Image(this.getClass().getResourceAsStream("/images/kacchan.png"));
+    private Image dukeImage = new Image(this.getClass().getResourceAsStream("/images/deku.png"));
+
+    @FXML
+    public void initialize() {
+        dialogContainer.getChildren().add(DialogBox.getDukeDialog(Ui.INTRO, dukeImage));
+        scrollPane.vvalueProperty().bind(dialogContainer.heightProperty());
+    }
+
+    public void setDuke(Deku d) {
+        deku = d;
+    }
+
+    /**
+     * Creates two dialog boxes, one echoing user input and the other containing Duke's reply and then appends them to
+     * the dialog container. Clears the user input after processing.
+     */
+    @FXML
+    private void handleUserInput() throws IOException, DekuException {
+        String input = userInput.getText();
+        String response = deku.getResponse(input);
+        dialogContainer.getChildren().addAll(
+                DialogBox.getUserDialog(input, userImage),
+                DialogBox.getDukeDialog(response, dukeImage)
+        );
+        userInput.clear();
+        if (this.deku.isClosed()) {
+            Platform.exit();
+        }
+        
+    }
+}
