+++ src/main/java/seedu/deku/Task.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.deku;
 
 /**
  * Represents a Task.
