+++ src/main/java/seedu/deku/ToDo.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.deku;
 
 /**
  * Represents a ToDo task.
