--- src/main/java/seedu/duke/DukeException.java
@@ -1,16 +1,0 @@
-package seedu.duke;
-
-public class DukeException extends Exception {
-
-    public DukeException(String message) {
-        super(message);
-    }
-
-
-
-//    public static void main(String[] args) {
-//        try {
-//
-//        }
-//    }
-}
